Separate out daemon tests into a module
